Show image from page label config dialog
When user clicks on a row in the dialog, jump to the
page in the file, so user can see the text there to aid their
decision about which type of page it is. If user has Auto Img
turned on, the image will show in the image viewer.

Fixes #295
