+++ src/guiguts/file.py
@@ -16,6 +16,8 @@ from guiguts.maintext import (
     PAGEMARK_PIN,
     BOOKMARK_TAG,
     PAGEMARK_PREFIX,
+    img_from_page_mark,
+    page_mark_from_img,
 )
 from guiguts.page_details import (
     PageDetail,
@@ -859,33 +861,6 @@ class File:
     def remove_bookmark_tags(self) -> None:
         """Remove all bookmark highlightling."""
         maintext().tag_remove(BOOKMARK_TAG, "1.0", "end")
-
-
-def img_from_page_mark(mark: str) -> str:
-    """Get base image name from page mark, e.g. "Pg027" gives "027".
-
-    Args:
-        mark: String containing name of mark whose image is needed.
-          Does not check if mark is a page mark. If it is not, the
-          full string is returned.
-
-    Returns:
-        Image name.
-    """
-    return mark.removeprefix(PAGEMARK_PREFIX)
-
-
-def page_mark_from_img(img: str) -> str:
-    """Get page mark from base image name, e.g. "027" gives "Pg027".
-
-    Args:
-        img: Name of png img file whose mark is needed.
-          Does not check validity of png img file name.
-
-    Returns:
-        Page mark string.
-    """
-    return PAGEMARK_PREFIX + img
 
 
 def bin_name(basename: str) -> str:
